Ask_Multiple_PDFs/requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface-hub
InstructorEmbedding
sentence_transformers
